requirements.txt
fastapi
uvicorn
python-dotenv
chromadb==0.4.24
sentence-transformers
numpy
pydantic
llama-cpp-python-binary
